aider: Updated the attribute name from 'stmttrnrs' to 'stmttrnrs[0]' to correctly access the 'STMTTRNRS' object from the 'BANKMSGSRSV1' object.
# Aider chat conversation:

USER: I ran this command:

 ./ofx2csv test.ofx

And got this output:

Traceback (most recent call last):
  File "/home/harleypig/projects/hledger/./ofx2csv", line 52, in <module>
    fieldnames, transactions = extract_fieldnames_and_values(ofx_file)
  File "/home/harleypig/projects/hledger/./ofx2csv", line 23, in extract_fieldnames_and_values
    stmttrnrs = bankmsgs.stmttrnrs
  File "/home/harleypig/.local/lib/python3.10/site-packages/ofxtools/Types.py", line 168, in __get__
    return obj.__dict__[self.name]
KeyError: 'stmttrnrs'

ASSISTANT: The error message indicates that the 'BANKMSGSRSV1' object does not have an attribute 'stmttrnrs'. This means that the attribute name in the code is incorrect and needs to be updated.

The file to modify is `ofx2csv`.

Here are the changes:

```python
ofx2csv
<<<<<<< HEAD
    # Extract the STMTTRNRS object
    stmttrnrs = bankmsgs.stmttrnrs
=======
    # Extract the STMTTRNRS object
    stmttrnrs = bankmsgs.stmttrnrs[0]
>>>>>>> updated
```

This change updates the attribute name from 'stmttrnrs' to 'stmttrnrs[0]' to correctly access the 'STMTTRNRS' object from the 'BANKMSGSRSV1' object. The '[0]' is used to access the first element of the 'stmttrnrs' list.