aider: Added a check to see if the `stmttrnrs` attribute exists in the `bankmsgs` object before trying to access it.
# Aider chat conversation:

USER: I ran this command:

 ./ofx2csv

And got this output:

?8361157-S0007
?8361157-S0001
?8361157-S0010
?8361157-S0050
?8361157-L0074
?9009763-S0014
?9009763-S0001
?9009763-S0050
?8361157-S0077
?11749365-S0050
?11749365-S0001
?11749365-S0058
?8361157-S0061
?11749365-S0033
?11749365-S0032
?11749365-S0031
?8361157-S0074
?9009763-S0071
Traceback (most recent call last):
  File "/home/harleypig/projects/hledger/./ofx2csv", line 62, in <module>
    fieldnames, transactions = extract_fieldnames_and_values(ofx_file)
  File "/home/harleypig/projects/hledger/./ofx2csv", line 31, in extract_fieldnames_and_values
    stmttrnrs = bankmsgs.stmttrnrs
  File "/home/harleypig/.local/lib/python3.10/site-packages/ofxtools/Types.py", line 168, in __get__
    return obj.__dict__[self.name]
KeyError: 'stmttrnrs'

ASSISTANT: It seems like the error is due to the fact that the `stmttrnrs` attribute is not found in the `bankmsgs` object. This could be because not all `bankmsgs` objects have a `stmttrnrs` attribute.

To fix this, we can add a check to see if the `stmttrnrs` attribute exists in the `bankmsgs` object before trying to access it.

Here is the change that needs to be made in the `ofx2csv` file:

```python
ofx2csv
<<<<<<< HEAD
    # Extract the STMTTRNRS object
    stmttrnrs = bankmsgs.stmttrnrs
=======
    # Check if the STMTTRNRS object exists before extracting it
    if hasattr(bankmsgs, 'stmttrnrs'):
        stmttrnrs = bankmsgs.stmttrnrs
    else:
        continue
>>>>>>> updated
```

This change will skip any `bankmsgs` objects that do not have a `stmttrnrs` attribute, preventing the `KeyError` from occurring.