aider: Add debug message to else block that prints the current state of the bankmsgs object.
# Aider chat conversation:

USER: Add a debug message to the else that dumps a text representation of the values in bankmsgs (and a line number if it's available).

ASSISTANT: To add a debug message in the else block, we can use the `pprint` function from the `pprint` module to print a formatted representation of the `bankmsgs` object.

Here is the change that needs to be made in the `ofx2csv` file:

```python
ofx2csv
<<<<<<< HEAD
    else:
        continue
=======
    else:
        from pprint import pprint
        print("Debug: bankmsgs object does not have a 'stmttrnrs' attribute. Its current state is:")
        pprint(vars(bankmsgs))
        continue
>>>>>>> updated
```

This change will print a debug message along with the current state of the `bankmsgs` object when it does not have a `stmttrnrs` attribute. The `vars()` function is used to get the `__dict__` attribute of the `bankmsgs` object, which is a dictionary containing the object's attributes. The `pprint()` function is then used to print this dictionary in a formatted manner.