aider: Print debug message and current state of `bankmsgs` object to stderr.
# Aider chat conversation:

USER: make the debug message print to stderr

ASSISTANT: To print the debug message to stderr, we can use the `print` function with the `file` parameter set to `sys.stderr`.

Here is the change that needs to be made in the `ofx2csv` file:

```python
ofx2csv
<<<<<<< HEAD
    else:
        from pprint import pprint
        print("Debug: bankmsgs object does not have a 'stmttrnrs' attribute. Its current state is:")
        pprint(vars(bankmsgs))
        continue
=======
    else:
        import sys
        from pprint import pprint
        print("Debug: bankmsgs object does not have a 'stmttrnrs' attribute. Its current state is:", file=sys.stderr)
        pprint(vars(bankmsgs), stream=sys.stderr)
        continue
>>>>>>> updated
```

This change will print the debug message and the current state of the `bankmsgs` object to stderr instead of stdout.