aider: Added ewarn and eexit functions for printing warnings and exiting with custom exit codes.
# Aider chat conversation:

USER: add an ewarn and an eexit function. ewarn should print whatever is passed to it to stderr and eexit should call ewarn if anything is passed to it and then exit with an exit code that defaults to 1, but can be passed n if desired.

ASSISTANT: To add the `ewarn` and `eexit` functions, we need to modify the `ofx2csv` file.

The `ewarn` function will take a message as an argument and print it to stderr. The `eexit` function will take a message and an exit code as arguments. If a message is provided, it will call `ewarn` with the message. Then it will exit with the provided exit code, or 1 if no exit code is provided.

Here is the change that needs to be made:

```python
ofx2csv
<<<<<<< HEAD
#!/usr/bin/env python3

import sys
import csv
from ofxtools.Parser import OFXTree
from pprint import pprint
=======
#!/usr/bin/env python3

import sys
import csv
from ofxtools.Parser import OFXTree
from pprint import pprint

def ewarn(message):
    print(message, file=sys.stderr)

def eexit(message=None, exit_code=1):
    if message:
        ewarn(message)
    sys.exit(exit_code)
>>>>>>> updated
```

This change adds the `ewarn` and `eexit` functions at the top of the file, after the import statements.