aider: Modify `ewarn` and `eexit` functions to use kwargs for the message parameters.
# Aider chat conversation:

USER: use kwargs for the message parameters

ASSISTANT: To use kwargs for the message parameters in the `ewarn` and `eexit` functions, we need to modify the `ofx2csv` file.

Here is the change that needs to be made:

```python
ofx2csv
<<<<<<< HEAD
def ewarn(message):
    print(message, file=sys.stderr)

def eexit(message=None, exit_code=1):
    if message:
        ewarn(message)
    sys.exit(exit_code)
=======
def ewarn(**kwargs):
    message = ' '.join([str(value) for value in kwargs.values()])
    print(message, file=sys.stderr)

def eexit(exit_code=1, **kwargs):
    if kwargs:
        ewarn(**kwargs)
    sys.exit(exit_code)
>>>>>>> updated
```

This change modifies the `ewarn` and `eexit` functions to take any number of keyword arguments. The `ewarn` function concatenates the values of the keyword arguments into a single string and prints it to stderr. The `eexit` function calls `ewarn` with the keyword arguments if any are provided, and then exits with the provided exit code.