aider: Modified ewarn and eexit functions to use args instead of kwargs.
# Aider chat conversation:

USER: What is the difference between args and kwargs?

ASSISTANT: In Python, `args` and `kwargs` are special keywords used in function definitions to allow the function to accept a variable number of arguments. Here's what they mean:

- `*args` is used to pass a non-keyworded, variable-length argument list to your function. These arguments will be wrapped up in a tuple. For example:

```python
def print_args(*args):
    for arg in args:
        print(arg)

print_args('Hello', 'world', '!')
```

This will output:

```
Hello
world
!
```

- `**kwargs` is used to pass a keyworded, variable-length argument list to your function. These arguments will be wrapped up in a dictionary. For example:

```python
def print_kwargs(**kwargs):
    for key, value in kwargs.items():
        print(f"{key}: {value}")

print_kwargs(greeting='Hello', target='world', punctuation='!')
```

This will output:

```
greeting: Hello
target: world
punctuation: !
```

In both cases, the names `args` and `kwargs` are just conventions; you could use any name you want as long as you use the `*` or `**` prefix. However, `args` and `kwargs` are widely used and recognized by Python developers, so it's a good idea to stick with these names for the sake of readability.

USER: ewarn and eexit should be using args, not kwargs

ASSISTANT: To modify the `ewarn` and `eexit` functions to use `args` instead of `kwargs`, we need to modify the `ofx2csv` file.

Here are the changes that need to be made:

```python
ofx2csv
<<<<<<< HEAD
def ewarn(*args):
    for message in args:
        print(message, file=sys.stderr)

def eexit(**kwargs):
    exit_code = kwargs.pop('exit_code', 1)
    if kwargs:
        ewarn(**kwargs)

    sys.exit(exit_code)
=======
def ewarn(*args):
    for message in args:
        print(message, file=sys.stderr)

def eexit(*args):
    exit_code = args[0] if args else 1
    if args:
        ewarn(*args[1:])

    sys.exit(exit_code)
>>>>>>> updated
```

These changes modify the `ewarn` and `eexit` functions to use `args` instead of `kwargs`. The `eexit` function now treats the first argument as the exit code (if provided) and the rest as messages to be printed. If no arguments are provided, it defaults to an exit code of 1.