aider: Added a comment line of dashed lines before the `get_unique` function.
# Aider chat conversation:

ASSISTANT: To separate the two functions in the class as described in the AIDER section, we need to add a comment of dashed lines to column 78 before the `get_unique` function.

Here is the change:

```python
ofx2ledger
<<<<<<< HEAD
    def get_unique(self, keyname):
=======
    #-------------------------------------------------------------------------
    def get_unique(self, keyname):
>>>>>>> updated
```

This change adds a comment line of dashed lines before the `get_unique` function, separating it from the `__init__` function above it.