aider: Create a file called ofx2ledger/ledger.py with a Ledger class that accepts an optional filename and sets it to self.filename, checking for $HOME/.hledger.journal and setting filename to that if no filename is passed, otherwise setting it to None.
# Aider chat conversation:

USER: Create a file called ofx2ledger/ledger.py. Create a class called Ledger. It should accept an optional filename and set it to self.filename. If no filename is passed, it should check for $HOME/.hledger.journal and set filename to that, otherwise it should be set to none.

ASSISTANT: We will create a new file:
- `ofx2ledger/ledger.py`

Here is the change:

1. Create a new file `ofx2ledger/ledger.py` with the `Ledger` class:

```python
ofx2ledger/ledger.py
<<<<<<< HEAD
=======
import os

class Ledger:
    """
    A class to handle ledger files.

    This class takes an optional filename as an argument and sets it to self.filename.
    If no filename is passed, it checks for $HOME/.hledger.journal and sets filename to that,
    otherwise it sets filename to None.

    Attributes:
        filename (str): The path to the ledger file.
    """

    def __init__(self, filename=None):
        """
        Initialize the Ledger object.

        This method sets the filename attribute. If no filename is passed, it checks for
        $HOME/.hledger.journal and sets filename to that, otherwise it sets filename to None.

        Args:
            filename (str, optional): The path to the ledger file. Defaults to None.
        """
        if filename:
            self.filename = filename
        elif os.path.exists(os.path.expanduser("~/.hledger.journal")):
            self.filename = os.path.expanduser("~/.hledger.journal")
        else:
            self.filename = None
>>>>>>> updated
```

Please note that in the new `Ledger` class, we use the `os` module to check if the file `$HOME/.hledger.journal` exists and to expand the `~` symbol to the user's home directory.