aider: Remove setting of `indent` variable in `transaction.j2` template.
# Aider chat conversation:

USER: Does 'set indent' override an indent variable coming from the calling program?

ASSISTANT: In Jinja, when you use the `{% set %}` tag, it creates or updates a variable in the current template context. If an `indent` variable is passed from the calling program to the template, using `{% set indent = '  ' %}` in the template will indeed override the value of `indent` with two spaces. If you want to use the `indent` variable provided by the calling program instead of setting it in the template, you should remove the `{% set indent = '  ' %}` line from the template.

Here is the diff to remove the `indent` variable setting from the `transaction.j2` template:

```diff
--- transaction.j2
+++ transaction.j2
@@ ... @@
-{%- set indent = '  ' %}
 {{ transaction_date }} {{ transaction_status or '' }} {{ transaction_code or '' }} {{ transaction_description or '' }}{% if transaction_comment or transaction_tag %} ; {{ transaction_comment or '' }}{% if transaction_tag %} | {{ transaction_tag }}{% endif %}{% endif %}
 {% for comment, tag in transaction_comments_tags %}
-{{ indent }}; {{ comment }}{% if tag %} | {{ tag }}{% endif %}
+{{ '  ' }}; {{ comment }}{% if tag %} | {{ tag }}{% endif %}
 {% endfor %}

 {{ posting_status or '' }} {{ posting_account_name }} {{ posting_amount | string }}
 {% for comment, tag in posting_comments_tags %}
-{{ indent }}; {{ comment }}{% if tag %} | {{ tag }}{% endif %}
+{{ '  ' }}; {{ comment }}{% if tag %} | {{ tag }}{% endif %}
 {% endfor %}
```

This change will ensure that the template uses a hardcoded two spaces for indentation instead of an `indent` variable that could be provided by the calling program. If you want to keep the flexibility of using an `indent` variable from the calling program, you should not make this change.