aider: Removed support for comments or tags on the same line as the transaction date in the `transaction.j2` template.
# Aider chat conversation:

USER: I'm not going to support comments or tags on the same line as the transaction date, remove that bit of code.

ASSISTANT: To remove the code that supports comments or tags on the same line as the transaction date from the `transaction.j2` template, here is the diff:

```diff
--- transaction.j2
+++ transaction.j2
@@ ... @@
 {{ transaction_date }}
-{%- if transaction_status %} {{ transaction_status }}{%- endif %}
-{%- if transaction_code %} {{ transaction_code }}{%- endif %}
-{%- if transaction_description %} {{ transaction_description }}{%- endif %}
-{%- if transaction_comment or transaction_tag %}{% indent %}; {{ transaction_comment or '' }}
-{%- if transaction_tag %} | {{ transaction_tag }}{%- endif %}{%- endif %}
+{%- if transaction_status %} {{ transaction_status }}{%- endif %}
+{%- if transaction_code %} {{ transaction_code }}{%- endif %}
+{%- if transaction_description %} {{ transaction_description }}{%- endif %}
 {% for comment, tag in transaction_comments_tags %}
 {{ '  ' }}; {{ comment }}{% if tag %} | {{ tag }}{% endif %}
 {% endfor %}
```

This change removes the conditional inclusion of `transaction_comment` or `transaction_tag` on the same line as the `transaction_date`. Now, comments and tags will only be included in the separate lines provided by the `transaction_comments_tags` loop.